refactor(summary-notification): Notifierクラスをアプリケーション層へ移動
