+++ lambda/summary-notification/src/application/__init__.py
@@ -1,9 +1,10 @@
-"""Application レイヤー: ビジネスロジック"""
-
 from .message_formatter import format_summary_message
+from .notifier import INotifier, NotificationFailed
 from .summary_notification_service import SummaryNotificationService
 
 __all__ = [
     "format_summary_message",
     "SummaryNotificationService",
+    "INotifier",
+    "NotificationFailed",
 ]
