+++ lambda/summary-notification/src/domain/__init__.py
@@ -6,10 +6,8 @@ from shared.domain.asset_record import AssetRecord
 from .asset_record_reader import IAssetRecordReader
 from .exceptions import (
     AssetRetrievalFailed,
-    NotificationFailed,
     SummaryNotificationFailed,
 )
-from .notifier import INotifier
 from .ops_indicators import OpsIndicators
 from .ops_indicators_service import calculate_ops_indicators
 
@@ -22,9 +20,7 @@ __all__ = [
     "calculate_ops_indicators",
     # Interfaces
     "IAssetRecordReader",
-    "INotifier",
     # Exceptions
     "SummaryNotificationFailed",
     "AssetRetrievalFailed",
-    "NotificationFailed",
 ]
