+++ lambda/summary-notification/src/domain/exceptions.py
@@ -1,5 +1,3 @@
-"""サマリ通知機能のカスタム例外定義"""
-
 from typing import Self
 
 
@@ -9,6 +7,7 @@ class SummaryNotificationFailed(Exception):
     pass
 
 
+# TODO: Repositoryと同様のファイルに移動する
 class AssetRetrievalFailed(SummaryNotificationFailed):
     """資産情報の取得エラー"""
 
@@ -29,25 +28,3 @@ class AssetRetrievalFailed(SummaryNotificationFailed):
             AssetRetrievalFailed: 生成された例外インスタンス
         """
         return cls("資産情報の取得中にエラーが発生しました")
-
-
-class NotificationFailed(SummaryNotificationFailed):
-    """通知送信エラー"""
-
-    @classmethod
-    def during_request(cls) -> Self:
-        """通知送信中にエラーが発生した場合の例外インスタンスを生成する名前付きコンストラクタ
-
-        Returns:
-            NotificationFailed: 生成された例外インスタンス
-        """
-        return cls("通知送信中にエラーが発生しました")
-
-    @classmethod
-    def before_request(cls) -> Self:
-        """通知送信前にエラーが発生した場合の例外インスタンスを生成する名前付きコンストラクタ
-
-        Returns:
-            NotificationFailed: 生成された例外インスタンス
-        """
-        return cls("通知送信前にエラーが発生しました")
